Move security config to new dir
