+++ hub/src/main/java/org/tapsensors/hub/config/security/BaseSecurityConfig.java
@@ -1,4 +1,4 @@
-package org.tapsensors.hub.config;
+package org.tapsensors.hub.config.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
